+++ frontend/app-development/config/routes.tsx
@@ -1,7 +1,7 @@
 import { SubApp } from '../../packages/ux-editor/src/SubApp';
 import { AccessControlContainer } from '../features/accessControl/containers/AccessControlContainer';
-import { About } from '../features/administration/components/About';
 import { Administration } from '../features/administration/components/Administration';
+import { LegacyAdministration } from '../features/administration/components/LegacyAdministration';
 import { TextEditor } from '../features/textEditor/TextEditor';
 import DataModellingContainer from '../features/dataModelling/containers/DataModellingContainer';
 import { TopBarMenu } from '../layout/AppBar/appBarConfig';
@@ -47,7 +47,7 @@ export const routes: IRoute[] = [
     activeSubHeaderSelection: TopBarMenu.About,
     activeLeftMenuSelection: 'Om appen',
     menu: 'about',
-    subapp: shouldDisplayFeature('newAbout') ? About : Administration,
+    subapp: shouldDisplayFeature('newAdministration') ? Administration : LegacyAdministration,
   },
   {
     path: '/:org/:app/datamodel',
